Factored more codes out to SolutionController shared examples.
